+++ services/jobrunner/Dockerfile
@@ -1,14 +1,14 @@
 # syntax=docker/dockerfile:1
 
-# Build stage: install scheduler extras using cached wheels
+# Build stage: install jobrunner extras using cached wheels
 FROM sidetrack/base:dev AS builder
 RUN apt-get update && apt-get install -y --no-install-recommends build-essential \
     && rm -rf /var/lib/apt/lists/*
-RUN pip install --no-cache-dir --no-index --find-links /wheels 'sidetrack[scheduler]'
+RUN pip install --no-cache-dir --no-index --find-links /wheels 'sidetrack[jobrunner,worker]'
 
 # Runtime stage: copy installed packages into base image
 FROM sidetrack/base:dev AS runtime
 COPY --from=builder /usr/local /usr/local
-RUN python -c "import sidetrack; print('scheduler runtime deps OK')"
+RUN python -c "import sidetrack; print('jobrunner runtime deps OK')"
 
-CMD ["python", "-m", "sidetrack.scheduler.run"]
+CMD ["python", "-m", "sidetrack.jobrunner.run"]
