--- services/scheduler/.gitkeep
@@ -1,1 +1,0 @@
-# Keep this directory in git
